refactor: make models module public
