+++ src/uipath/_services/assets_service.py
@@ -5,8 +5,8 @@ from httpx import Response
 from .._config import Config
 from .._execution_context import ExecutionContext
 from .._folder_context import FolderContext
-from .._models import UserAsset
 from .._utils import Endpoint, RequestSpec, header_folder, infer_bindings
+from ..models import UserAsset
 from ._base_service import BaseService
