+++ src/uipath/_services/context_grounding_service.py
@@ -6,14 +6,14 @@ from pydantic import TypeAdapter
 from .._config import Config
 from .._execution_context import ExecutionContext
 from .._folder_context import FolderContext
-from .._models import IngestionInProgressException
-from .._models.context_grounding import ContextGroundingQueryResponse
-from .._models.context_grounding_index import ContextGroundingIndex
 from .._utils import Endpoint, RequestSpec
 from .._utils.constants import (
     HEADER_FOLDER_KEY,
     ORCHESTRATOR_STORAGE_BUCKET_DATA_SOURCE,
 )
+from ..models import IngestionInProgressException
+from ..models.context_grounding import ContextGroundingQueryResponse
+from ..models.context_grounding_index import ContextGroundingIndex
 from ._base_service import BaseService
 from .folder_service import FolderService
