+++ src/uipath/_services/queues_service.py
@@ -5,8 +5,8 @@ from httpx import Response
 from .._config import Config
 from .._execution_context import ExecutionContext
 from .._folder_context import FolderContext
-from .._models import CommitType, QueueItem, TransactionItem, TransactionItemResult
 from .._utils import Endpoint, RequestSpec
+from ..models import CommitType, QueueItem, TransactionItem, TransactionItemResult
 from ._base_service import BaseService
