+++ src/uipath/models/assets.py
