+++ tests/sdk/services/test_uipath_llm_integration.py
@@ -6,17 +6,17 @@ import pytest
 
 from uipath._config import Config
 from uipath._execution_context import ExecutionContext
-from uipath._models.llm_gateway import (
+from uipath._services.llm_gateway_service import (
+    ChatModels,
+    UiPathLlmChatService,
+)
+from uipath.models.llm_gateway import (
     AutoToolChoice,
     SpecificToolChoice,
     ToolDefinition,
     ToolFunctionDefinition,
     ToolParametersDefinition,
     ToolPropertyDefinition,
-)
-from uipath._services.llm_gateway_service import (
-    ChatModels,
-    UiPathLlmChatService,
 )
